[core] Add nonblocking task cancellation
Add a `::tryCancel()` method to the WorkTask so objects
that own a WorkTask can try to cancel and if it will block,
postpone the cancellation.

On debug mode, the RunLoop will print a warning when we are
blocking trying to cancel tasks on the main thread which we
should never do.